+++ test/unit/fuzz-tools.test.js
@@ -264,6 +264,7 @@ import * as metabolicV273 from '../../lib/metabolic-v273.js';
 import * as proteinsV274 from '../../lib/proteins-v274.js';
 import * as fibrosisV275 from '../../lib/fibrosis-v275.js';
 import * as nutritionV276 from '../../lib/nutrition-v276.js';
+import * as renalV277 from '../../lib/renal-v277.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -472,6 +473,7 @@ const MODULES = {
   'proteins-v274.js': proteinsV274,
   'fibrosis-v275.js': fibrosisV275,
   'nutrition-v276.js': nutritionV276,
+  'renal-v277.js': renalV277,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];